client/fed-learn-client-agent/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
cryptography==43.0.3
fastapi==0.119.0
flatbuffers==25.9.23
flwr==1.15.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.15.1
idna==3.11
iterators==0.0.2
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mysql-connector-python==9.2.0
namex==0.1.0
numba==0.62.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
protobuf==4.25.8
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==13.9.4
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
starlette==0.48.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.3.0
tomli_w==1.2.0
tqdm==4.67.1
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
